refactor: move test code for product domain under the 'test/product' package
