+++ src/test/java/kitchenpos/Fixtures.java
@@ -16,6 +16,8 @@ import java.util.Arrays;
 import java.util.Random;
 import java.util.UUID;
 
+import static kitchenpos.product.ProductFixtures.product;
+
 public class Fixtures {
     public static final UUID INVALID_ID = new UUID(0L, 0L);
 
@@ -115,17 +117,5 @@ public class Fixtures {
         orderTable.setNumberOfGuests(numberOfGuests);
         orderTable.setOccupied(occupied);
         return orderTable;
-    }
-
-    public static Product product() {
-        return product("후라이드", 16_000L);
-    }
-
-    public static Product product(final String name, final long price) {
-        final Product product = new Product();
-        product.setId(UUID.randomUUID());
-        product.setName(name);
-        product.setPrice(BigDecimal.valueOf(price));
-        return product;
     }
 }
