+++ src/test/java/kitchenpos/asis/application/MenuServiceTest.java
@@ -7,6 +7,7 @@ import kitchenpos.asis.domain.MenuRepository;
 import kitchenpos.asis.domain.Product;
 import kitchenpos.asis.domain.ProductRepository;
 import kitchenpos.asis.infra.PurgomalumClient;
+import kitchenpos.product.InMemoryProductRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -27,7 +28,7 @@ import static kitchenpos.Fixtures.INVALID_ID;
 import static kitchenpos.Fixtures.menu;
 import static kitchenpos.Fixtures.menuGroup;
 import static kitchenpos.Fixtures.menuProduct;
-import static kitchenpos.Fixtures.product;
+import static kitchenpos.product.ProductFixtures.product;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
